requirements.txt
altair==6.0.0
asttokens==3.0.0
attrs==26.1.0
bleach==6.2.0
blinker==1.9.0
cachetools==7.0.5
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
filelock==3.25.2
fonttools==4.57.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.7.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.3
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==2.18.1
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.50
protobuf==6.30.2
psutil==7.0.0
pure_eval==0.2.3
pyarrow==23.0.1
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
pywin32==310
pyzmq==26.3.0
referencing==0.37.0
requests==2.32.3
rpds-py==0.30.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
smmap==5.0.3
stack-data==0.6.3
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
torch==2.11.0
torchvision==0.26.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
